Changed: move the category class and service inside the category feature
