+++ src/app/admin/categories/create.component.ts
@@ -1,7 +1,7 @@
 import { Component, Output, EventEmitter } from '@angular/core';
 
-import { CategoryService } from '../../services/category.service';
-import { Category } from '../../classes/category';
+import { CategoryService } from './category.service';
+import { Category } from './category';
 
 @Component({
     selector: 'xp-admin-category-create',
