+++ src/app/app.module.ts
@@ -21,7 +21,7 @@ import { AdminComponent } from './admin/admin.component';
 import { AdminGuard } from "./core/admin.guard";
 import { AdminCategoriesComponent } from './admin/categories/categories.component';
 import { AdminTasksComponent } from './admin/admin-tasks/admin-tasks.component';
-import { CategoryService } from "./services/category.service";
+import { CategoryService } from "./admin/categories/category.service";
 import { CategoryFormComponent } from './admin/categories/form/form.component';
 import { TaskService } from './services/task.service';
 import { AssessmentTypeService } from './services/assessment-type.service';
